fix: move async to editColumnTItleThunk
